+++ mulog-elasticsearch/src/com/brunobonacci/mulog/publishers/elasticsearch/util.clj
@@ -1,5 +1,6 @@
-(ns com.brunobonacci.mulog.publishers.util
+(ns com.brunobonacci.mulog.publishers.elasticsearch.util
   (:require [clojure.string :as str]
+            [clojure.walk :as w]
             [com.brunobonacci.mulog.utils :as ut]))
 
 
@@ -39,11 +40,26 @@
     (double? v)     [(str k ".f") v]
     (float? v)      [(str k ".f") v]
     (map? v)        [(str k ".o") v]
-    (sequential? v) [(str k ".a") v]
-    (set? v)        [(str k ".a") v]
+    (sequential? v) [(str k ".a") (mapv (fn [item] (conj {} (type-mangle ["_aVal" item]))) v)]
+    (set? v)        [(str k ".a") (mapv (fn [item] (conj {} (type-mangle ["_aVal" item]))) v)]
     (boolean? v)    [(str k ".b") v]
     (keyword? v)    [(str k ".k") v]
+    (symbol? v)     [(str k ".k") (str v)]
     (instance? java.util.Date v) [(str k ".t") v]
     (instance? Throwable v) [(str k ".x") (ut/exception-stacktrace v)]
     :else e
     ))
+
+
+
+(defn mangle-map
+  [m]
+  (let [mangler (comp type-mangle snake-case-mangle)]
+    (w/postwalk
+      (fn [i]
+        (if (map? i)
+          (->> i
+            (map (fn [entry] (mangler entry)))
+            (into {}))
+          i))
+      m)))